+++ src/evaldet/tracks.py
@@ -1,8 +1,9 @@
+import collections as co
 import csv
+import datetime as dt
+import pathlib
 import typing as t
 import xml.etree.ElementTree as ET
-from pathlib import Path
-from typing import Any, NamedTuple, Optional, Union
 
 import numpy as np
 
@@ -16,11 +17,14 @@ _WIDTH_KEY = "width"
 _HEIGHT_KEY = "height"
 
 
-class FrameTracks(NamedTuple):
+class FrameTracks(t.NamedTuple):
     detections: np.ndarray
     ids: np.ndarray
-    classes: Optional[np.ndarray]
-    confs: Optional[np.ndarray]
+    classes: np.ndarray
+    confs: np.ndarray
+
+
+TracksType = t.TypeVar("TracksType", bound="Tracks")
 
 
 class Tracks:
@@ -42,11 +46,11 @@ class Tracks:
 
     @classmethod
     def from_csv(
-        cls,
-        csv_file: Union[str, Path],
+        cls: t.Type[TracksType],
+        csv_file: t.Union[str, pathlib.Path],
         fieldnames: t.List[str],
         zero_indexed: bool = True,
-    ) -> "Tracks":
+    ) -> TracksType:
         """Get detections from a CSV file.
 
         The CSV file should have a normal comma (,) as a separator, and should not
@@ -73,7 +77,7 @@ class Tracks:
         tracks = cls()
         with open(csv_file, newline="") as file:
             csv_reader = csv.DictReader(file, fieldnames=fieldnames, dialect="unix")
-            frames: t.Dict[int, Any] = {}
+            frames: t.Dict[int, t.Any] = {}
 
             for line_num, line in enumerate(csv_reader):
                 try:
@@ -109,7 +113,9 @@ class Tracks:
         return tracks
 
     @classmethod
-    def from_mot(cls, file_path: Union[Path, str]) -> "Tracks":
+    def from_mot(
+        cls: t.Type[TracksType], file_path: t.Union[pathlib.Path, str]
+    ) -> TracksType:
         """Creates a Tracks object from detections file in the MOT format.
 
         The format should look like this::
@@ -141,7 +147,9 @@ class Tracks:
         return cls.from_csv(file_path, fieldnames, zero_indexed=False)
 
     @classmethod
-    def from_mot_gt(cls, file_path: Union[Path, str]) -> "Tracks":
+    def from_mot_gt(
+        cls: t.Type[TracksType], file_path: t.Union[pathlib.Path, str]
+    ) -> TracksType:
         """Creates a Tracks object from detections file in the MOT ground truth format.
         This format has some more information compared to the normal
 
@@ -175,7 +183,9 @@ class Tracks:
         return cls.from_csv(file_path, fieldnames, zero_indexed=False)
 
     @classmethod
-    def from_mot_cvat(cls, file_path: Union[Path, str]) -> "Tracks":
+    def from_mot_cvat(
+        cls: t.Type[TracksType], file_path: t.Union[pathlib.Path, str]
+    ) -> "Tracks":
         """Creates a Tracks object from detections file in the CVAT's MOT format.
 
         The format should look like this::
@@ -210,11 +220,11 @@ class Tracks:
 
     @classmethod
     def from_ua_detrac(
-        cls,
-        file_path: Union[Path, str],
-        classes_attr_name: Optional[str] = None,
-        classes_list: Optional[t.List[str]] = None,
-    ) -> "Tracks":
+        cls: t.Type[TracksType],
+        file_path: t.Union[pathlib.Path, str],
+        classes_attr_name: t.Optional[str] = None,
+        classes_list: t.Optional[t.List[str]] = None,
+    ) -> TracksType:
         """Creates a Tracks object from detections file in the UA-DETRAC XML format.
 
         Here's how this file might look like:
@@ -314,10 +324,10 @@ class Tracks:
 
     @classmethod
     def from_cvat_video(
-        cls,
-        file_path: Union[Path, str],
+        cls: t.Type[TracksType],
+        file_path: t.Union[pathlib.Path, str],
         classes_list: t.List[str],
-    ) -> "Tracks":
+    ) -> TracksType:
         """Creates a Tracks object from detections file in the CVAT for Video XML
         format.
 
@@ -367,7 +377,9 @@ class Tracks:
         root = xml_tree.getroot()
         tracks = cls()
 
-        frames: t.Dict[int, Any] = {}
+        frames: t.Dict[int, t.Dict[str, t.List]] = co.defaultdict(
+            lambda: co.defaultdict(list)
+        )
         tracks_cvat = root.findall("track")
         for track_cvat in tracks_cvat:
             track_id = int(track_cvat.attrib[_ID_KEY])
@@ -382,16 +394,10 @@ class Tracks:
                 width = float(box.attrib["xbr"]) - xmin
                 height = float(box.attrib["ybr"]) - ymin
 
-                current_frame = frames.get(frame_num, {})
-                if current_frame:
-                    current_frame["classes"].append(track_class)
-                    current_frame["ids"].append(track_id)
-                    current_frame["detections"].append([xmin, ymin, width, height])
-                else:
-                    current_frame["classes"] = [track_class]
-                    current_frame["ids"] = [track_id]
-                    current_frame["detections"] = [[xmin, ymin, width, height]]
-                    frames[frame_num] = current_frame
+                current_frame = frames[frame_num]
+                current_frame["classes"].append(track_class)
+                current_frame["ids"].append(track_id)
+                current_frame["detections"].append([xmin, ymin, width, height])
 
         list_frames = sorted(list(frames.keys()))
         for frame in list_frames:
@@ -439,21 +445,22 @@ class Tracks:
 
     def __init__(self) -> None:
 
-        self._frame_nums: t.Set[int] = set()
+        self._frames: t.Set[int] = set()
         self._detections: t.Dict[int, np.ndarray] = dict()
         self._ids: t.Dict[int, np.ndarray] = dict()
         self._classes: t.Dict[int, np.ndarray] = dict()
         self._confs: t.Dict[int, np.ndarray] = dict()
+        self._id_to_frames: t.Dict[int, t.Set[int]] = co.defaultdict(set)
 
     def add_frame(
         self,
         frame_num: int,
-        ids: Union[t.List[int], np.ndarray],
+        ids: t.Union[t.List[int], np.ndarray],
         detections: np.ndarray,
-        classes: Optional[Union[t.List[int], np.ndarray]] = None,
-        confs: Optional[Union[t.List[float], np.ndarray]] = None,
+        classes: t.Optional[t.Union[t.List[int], np.ndarray]] = None,
+        confs: t.Optional[t.Union[t.List[float], np.ndarray]] = None,
     ) -> None:
-        """Add a frame to the collection. Can overwrite existing frame.
+        """Add a frame to the collection. Can overwrite an existing frame.
 
         Args:
             frame_num: A non-negative frame number
@@ -461,10 +468,9 @@ class Tracks:
             detections: An Nx4 array describing the bounding boxes of objects in
                 the frame. It should be in the ``xywh`` format.
             classes: An optional list (or numpy array) of classes for the objects.
-                If passed all objects in the frame must be assigned a class.
+                If not passed all objects in the frame will get a class of 0.
             confs: An optional list (or numpy array) of confidence scores for the
-                objects. If passed all objects in the frame must be assigned a
-                confidence score.
+                objects. If not passed, all detection will get a confidence of 1.
         """
 
         if len(ids) == 0:
@@ -504,23 +510,23 @@ class Tracks:
             )
 
         # If all ok, add objects to collections
-        self._detections[frame_num] = detections.copy()
-
-        def _to_numpy(x: Union[list, np.ndarray], is_int: bool = True) -> np.ndarray:
-            if isinstance(x, list):
-                return np.array(x).astype(np.int64 if is_int else np.float64)
-            else:
-                return x.copy().astype(np.int64 if is_int else np.float64)
-
-        self._ids[frame_num] = _to_numpy(ids)
+        self._detections[frame_num] = np.array(detections, copy=True, dtype=np.float32)
+        self._ids[frame_num] = np.array(ids, copy=True, dtype=np.int32)
 
         if classes is not None:
-            self._classes[frame_num] = _to_numpy(classes)
+            self._classes[frame_num] = np.array(classes, copy=True, dtype=np.int32)
+        else:
+            self._classes[frame_num] = np.full((len(ids),), 0, dtype=np.int32)
 
         if confs is not None:
-            self._confs[frame_num] = _to_numpy(confs, is_int=False)
+            self._confs[frame_num] = np.array(confs, copy=True, dtype=np.float32)
+        else:
+            self._confs[frame_num] = np.full((len(ids),), 1, dtype=np.float32)
 
-        self._frame_nums.add(frame_num)
+        self._frames.add(frame_num)
+
+        for _id in ids:
+            self._id_to_frames[_id].add(frame_num)
 
     def filter_frame(self, frame_num: int, filter: np.ndarray) -> None:
         """Filters a frame and all its objects according to the filter array.
@@ -554,14 +560,16 @@ class Tracks:
         if filter.min() == 1:
             return
 
+        for id_out in frame.ids[~filter]:
+            if len(self._id_to_frames[id_out]) == 1:
+                del self._id_to_frames[id_out]
+            else:
+                self._id_to_frames[id_out].remove(frame_num)
+
         self._detections[frame_num] = frame.detections[filter]
         self._ids[frame_num] = frame.ids[filter]
-
-        if frame.confs is not None:
-            self._confs[frame_num] = frame.confs[filter]
-
-        if frame.classes is not None:
-            self._classes[frame_num] = frame.classes[filter]
+        self._confs[frame_num] = frame.confs[filter]
+        self._classes[frame_num] = frame.classes[filter]
 
     def filter_by_class(self, classes: t.List[int]) -> None:
         """Filter all frames by classes
@@ -573,15 +581,9 @@ class Tracks:
         Args:
             classes: A list of which class labels to keep.
         """
-        if not self._classes:
-            raise ValueError("Can not filter by class, no class data")
-
-        for frame in self._frame_nums.copy():
-            if frame not in self._classes:
-                del self[frame]
-            else:
-                filter_cls = np.in1d(self._classes[frame], classes)
-                self.filter_frame(frame, filter_cls)
+        for frame in self._frames.copy():
+            filter_cls = np.in1d(self._classes[frame], classes)
+            self.filter_frame(frame, filter_cls)
 
     def filter_by_conf(self, lower_bound: float) -> None:
         """Filter all frames by confidence
@@ -594,21 +596,15 @@ class Tracks:
             lower_bound: The lower bound on the confidence value of the items,
                 so that they are not filtered out.
         """
-        if not self._confs:
-            raise ValueError("Can not filter by confidence, no confidence data")
-
-        for frame in self._frame_nums.copy():
-            if frame not in self._confs:
-                del self[frame]
-            else:
-                filter_conf = self._confs[frame] >= lower_bound
-                self.filter_frame(frame, filter_conf)
+        for frame in self._frames.copy():
+            filter_conf = self._confs[frame] >= lower_bound
+            self.filter_frame(frame, filter_conf)
 
     @property
     def all_classes(self) -> t.Set[int]:
         """Get a set of all classes in the collection."""
         classes: t.Set[int] = set()
-        for frame in self._frame_nums:
+        for frame in self._frames:
             if frame in self._classes:
                 classes.update(self._classes[frame])
 
@@ -622,63 +618,320 @@ class Tracks:
             A dictionary where keys are the track ids, and values
             are the numbers of frames they appear in.
         """
-        ids_count: t.Dict[int, int] = dict()
-        for frame in self._frame_nums:
-            for _id in self._ids[frame]:
-                ids_count[_id] = ids_count.get(_id, 0) + 1
-
+        ids_count = {k: len(v) for k, v in self._id_to_frames.items()}
         return ids_count
 
     @property
     def frames(self) -> t.Set[int]:
         """Get an ordered list of all frame numbers in the collection."""
-        return self._frame_nums.copy()
+        return self._frames.copy()
+
+    @property
+    def id_to_frames(self) -> t.Dict[int, t.Set[int]]:
+        """Get an ordered list of all frame numbers in the collection."""
+        return {k: v.copy() for k, v in self._id_to_frames.items()}
 
     def __len__(self) -> int:
-        return len(self._frame_nums)
+        return len(self._frames)
 
     def __contains__(self, idx: int) -> bool:
         """Whether the frame ``idx`` is present in the collection."""
-        return idx in self._frame_nums
+        return idx in self._frames
 
+    @t.overload
     def __getitem__(self, idx: int) -> FrameTracks:
         """Get the frame with number ``idx``.
 
-        Returns:
-            A dictionary with the key ``'ids'``, ``'detections'`` and, if available,
-            ``'classes'`` and ``confs``. The values are numpy arrays, with values
-            for each item in the frame.
+        Note that indexing with negative values is not supported.
         """
-        if idx not in self:
-            raise KeyError(f"The frame {idx} does not exist.")
 
-        classes = None
-        if idx in self._classes:
-            classes = self._classes[idx]
+    @t.overload
+    def __getitem__(self, idx: slice) -> "Tracks":
+        """Select only a subset of frames, as defined by the slice.
 
-        confs = None
-        if idx in self._confs:
-            confs = self._confs[idx]
+        Note that the ``step`` argument is not supported and will result in an error
+        being raised if it is supplied. Negative indices for start or stop argument are
+        similarly not supported.
+        """
 
-        return FrameTracks(
-            ids=self._ids[idx],
-            detections=self._detections[idx],
-            classes=classes,
-            confs=confs,
-        )
+    def __getitem__(self, idx: t.Union[int, slice]) -> t.Union[FrameTracks, "Tracks"]:
+
+        if isinstance(idx, int):
+            if idx < 0:
+                raise ValueError("Indexing with negative values is not supported.")
+            if idx not in self:
+                raise KeyError(f"The frame {idx} does not exist.")
+
+            return FrameTracks(
+                ids=self._ids[idx],
+                detections=self._detections[idx],
+                classes=self._classes[idx],
+                confs=self._confs[idx],
+            )
+
+        elif isinstance(idx, slice):
+            start, stop, step = idx.start, idx.stop, idx.step
+            if step is not None:
+                raise ValueError("Slicing with the step argument is not supported.")
+            if start < 0 or stop < 0:
+                raise ValueError("Slicing with negative indices is not supported.")
+
+            keep_frames = set(list(range(start, stop)))
+            keep_frames = self._frames.intersection(keep_frames)
+
+            new_tracks = type(self)()
+
+            new_tracks._frames = keep_frames
+            new_tracks._detections.update(
+                {fnum: self._detections[fnum].copy() for fnum in keep_frames}
+            )
+            new_tracks._ids.update(
+                {fnum: self._ids[fnum].copy() for fnum in keep_frames}
+            )
+            new_tracks._classes.update(
+                {fnum: self._classes[fnum].copy() for fnum in keep_frames}
+            )
+            new_tracks._confs.update(
+                {fnum: self._confs[fnum].copy() for fnum in keep_frames}
+            )
+
+            new_id_to_frames = {
+                k: v.intersection(keep_frames) for k, v in self._id_to_frames.items()
+            }
+            new_tracks._id_to_frames.update(
+                {k: v for k, v in new_id_to_frames.items() if len(v) > 0}
+            )
+            return new_tracks
+        else:
+            raise ValueError("Unrecognized index type")
 
     def __delitem__(self, frame_num: int) -> None:
         """Remove the frame with number ``frame_num``"""
 
         if frame_num not in self:
-            raise KeyError(f"Tracks object does not contain frame {frame_num}")
+            raise KeyError(f"Tracks object does not contain frame {frame_num}.")
 
-        self._frame_nums.remove(frame_num)
+        for id_out in self._ids[frame_num]:
+            if len(self._id_to_frames[id_out]) == 1:
+                del self._id_to_frames[id_out]
+            else:
+                self._id_to_frames[id_out].remove(frame_num)
+
+        self._frames.remove(frame_num)
         del self._ids[frame_num]
         del self._detections[frame_num]
+        del self._classes[frame_num]
+        del self._confs[frame_num]
 
-        if frame_num in self._classes:
-            del self._classes[frame_num]
+    def to_cvat_video(
+        self,
+        filename: t.Union[pathlib.Path, str],
+        labels: t.Sequence[str],
+        image_size: t.Tuple[int, int] = (1, 1),
+    ) -> None:
+        """Export detections to CVAT for Video 1.1 format.
 
-        if frame_num in self._confs:
-            del self._confs[frame_num]
+        More information on the format can be found `here <https://opencv.github.io/cvat/docs/manual/advanced/xml_format/>`_.
+
+        Args:
+            filename: The name of the file to save to - should have an ``.xml`` suffix.
+            labels: A list/tuple of label names. The length should be at least the
+                maximum label index - 1 (the first label corresponds to label at the
+                0-th index).
+            image_size: The size of the image in the ``[w, h]`` format, in pixels.
+        """
+
+        if len(self._id_to_frames) == 0:
+            max_label_ind = 1
+        else:
+            max_label_ind = max(self.all_classes) + 1
+        if len(labels) < max_label_ind:
+            raise ValueError(
+                f"The length of provied labels {len(labels)} is less than the largest"
+                f" label id (+1) among tracklets {max_label_ind}."
+            )
+
+        w, h = image_size
+        if len(self._frames) == 0:
+            max_frame = 1
+        else:
+            max_frame = max(self._frames)
+
+        annotations = ET.Element("annotations")
+        version = ET.SubElement(annotations, "version")
+        version.text = "1.1"
+
+        meta = ET.SubElement(annotations, "meta")
+
+        dumped = ET.SubElement(meta, "dumped")
+        now = dt.datetime.now().replace(tzinfo=None)
+        dumped.text = now.isoformat().replace("T", " ")
+
+        task = ET.SubElement(meta, "task")
+        task_id = ET.SubElement(task, "id")
+        task_id.text = "1"
+        task_name = ET.SubElement(task, "name")
+        task_name.text = "Tracking"
+        task_size = ET.SubElement(task, "size")
+        task_size.text = str(max_frame)
+        task_mode = ET.SubElement(task, "mode")
+        task_mode.text = "interpolation"
+        ET.SubElement(task, "overlap")
+        ET.SubElement(task, "bugtracker")
+        task_flipped = ET.SubElement(task, "flipped")
+        task_flipped.text = "False"
+        task_created = ET.SubElement(task, "created")
+        task_created.text = dumped.text
+        task_updated = ET.SubElement(task, "updated")
+        task_updated.text = dumped.text
+
+        labels_el = ET.SubElement(task, "labels")
+        for label_name in labels:
+            label_el = ET.SubElement(labels_el, "label")
+            label_name_el = ET.SubElement(label_el, "name")
+            label_name_el.text = label_name
+            ET.SubElement(label_el, "attributes")
+
+        owner = ET.SubElement(task, "owner")
+        ET.SubElement(owner, "username")
+        ET.SubElement(owner, "email")
+
+        original_size = ET.SubElement(task, "original_size")
+        original_size_w = ET.SubElement(original_size, "width")
+        original_size_h = ET.SubElement(original_size, "height")
+        original_size_w.text = f"{w}"
+        original_size_h.text = f"{h}"
+
+        segments = ET.SubElement(task, "segments")
+        segment = ET.SubElement(segments, "segment")
+        segment_id = ET.SubElement(segment, "id")
+        segment_id.text = "0"
+        segment_start = ET.SubElement(segment, "start")
+        segment_start.text = "0"
+        segment_stop = ET.SubElement(segment, "stop")
+        segment_stop.text = str(max_frame)
+        ET.SubElement(segment, "url")
+
+        for track_id, track_frames_set in self._id_to_frames.items():
+            track_frames = sorted(list(track_frames_set))
+
+            # Take first label, as only one is supported
+            id_first_ind = np.nonzero(self._ids[track_frames[0]] == track_id)[0][0]
+            label: int = self._classes[track_frames[0]][id_first_ind]
+
+            track_el = ET.SubElement(
+                annotations,
+                "track",
+                id=str(track_id),
+                label=labels[label],
+                source=type(self).__name__,
+            )
+
+            for i, frame_num in enumerate(track_frames):
+                id_frame_ind = np.nonzero(self._ids[frame_num] == track_id)[0][0]
+                bbox = self._detections[frame_num][id_frame_ind]
+                ET.SubElement(
+                    track_el,
+                    "box",
+                    frame=str(frame_num),
+                    xtl=f"{bbox[0]:.2f}",
+                    ytl=f"{bbox[1]:.2f}",
+                    xbr=f"{bbox[0] + bbox[2]:.2f}",
+                    ybr=f"{bbox[1] + bbox[3]:.2f}",
+                    outside="0",
+                    occluded="0",
+                    keyframe="1",
+                )
+
+                # Add fake element with outside=1 to prevent it showing up on CVAT
+                if (i + 1) == len(track_frames) or (frame_num + 1) < track_frames[
+                    i + 1
+                ]:
+                    ET.SubElement(
+                        track_el,
+                        "box",
+                        frame=str(frame_num + 1),
+                        xtl=f"{bbox[0]:.2f}",
+                        ytl=f"{bbox[1]:.2f}",
+                        xbr=f"{bbox[0] + bbox[2]:.2f}",
+                        ybr=f"{bbox[1] + bbox[3]:.2f}",
+                        outside="1",  # Make it outside to prevent ghost frames
+                        occluded="0",
+                        keyframe="1",
+                    )
+
+        tree = ET.ElementTree(annotations)
+        tree.write(
+            filename, xml_declaration=True, short_empty_elements=False, encoding="utf-8"
+        )
+
+    def to_csv(
+        self,
+        dirname: t.Union[pathlib.Path, str],
+        labels: t.Sequence[str],
+    ) -> None:
+        """Export detections to a simple CSV format. The format comprises of two files:
+        ``dets.csv``, containing the detections, and ``labels.txt``, which contains
+        the names of the labels (corresponding to label indices in ``dets.csv``). The
+        rows in ``dets.csv`` have the following format:
+
+            <frame>, <id>, <x_min>, <y_min>, <width>, <height>, <class>, <conf>
+
+        Note that ``frame`` and ``class`` are both 0 indexed.
+
+        Args:
+            dirname: The name of the directory to save to - will be created if it
+                doesn't already exist.
+            labels: A list/tuple of label names. The length should be at least the
+                maximum label index - 1 (the first label corresponds to label at the
+                0-th index).
+        """
+
+        if len(self._id_to_frames) == 0:
+            max_label_ind = 1
+        else:
+            max_label_ind = max(self.all_classes) + 1
+        if len(labels) < max_label_ind:
+            raise ValueError(
+                f"The length of provied labels {len(labels)} is less than the largest"
+                f" label id (+1) among tracklets {max_label_ind}."
+            )
+
+        # Create directory and write labels file
+        pathlib.Path(dirname).mkdir(parents=True, exist_ok=True)
+        with open(pathlib.Path(dirname) / "labels.txt", "w") as f:
+            for label in labels:
+                f.write(label + "\n")
+
+        # Export detections
+        fieldnames = [
+            "frame_id",
+            "track_id",
+            "x_min",
+            "y_min",
+            "w",
+            "h",
+            "class_id",
+            "conf",
+        ]
+        with open(pathlib.Path(dirname) / "dets.csv", "w", newline="") as f:
+            writer = csv.DictWriter(f, fieldnames=fieldnames)
+
+            for frame in sorted(self._frames):
+                dets = self._detections[frame]
+                ids = self._ids[frame]
+                classes = self._classes[frame]
+                confs = self._confs[frame]
+
+                for det_idx in range(len(dets)):
+                    item: t.Dict[str, t.Union[str, int, float]] = {
+                        "frame_id": frame,
+                        "track_id": ids[det_idx],
+                        "conf": confs[det_idx],
+                        "class_id": classes[det_idx],
+                        "x_min": f"{dets[det_idx][0]:.2f}",
+                        "y_min": f"{dets[det_idx][1]:.2f}",
+                        "w": f"{dets[det_idx][2]:.2f}",
+                        "h": f"{dets[det_idx][3]:.2f}",
+                    }
+                    writer.writerow(item)