+++ tests/conftest.py
@@ -1,0 +1,8 @@
+import pathlib
+
+import pytest
+
+
+@pytest.fixture
+def data_dir() -> pathlib.Path:
+    return pathlib.Path(__file__).parent / "data"